+++ src/autostore/types/__init__.py
@@ -1,5 +1,6 @@
 """Types."""
 
+from .fields import Role
 from .sqlalchemy import FloatArrayTypeDecorator, PathTypeDecorator
 
-__all__ = ["FloatArrayTypeDecorator", "PathTypeDecorator"]
+__all__ = ["Role", "FloatArrayTypeDecorator", "PathTypeDecorator"]